Move ServerDetails and other structs to pkg/model
