+++ internal/api/handlers/v0/publish_test.go
@@ -97,7 +97,7 @@ func TestPublishEndpoint(t *testing.T) {
 						Source: "github",
 						ID:     "example/test-server",
 					},
-					VersionDetail: model.VersionDetail{
+					VersionDetail: pkgmodel.VersionDetail{
 						Version: "1.0.0",
 					},
 				},
@@ -125,7 +125,7 @@ func TestPublishEndpoint(t *testing.T) {
 						Source: "github",
 						ID:     "example/test-server",
 					},
-					VersionDetail: model.VersionDetail{
+					VersionDetail: pkgmodel.VersionDetail{
 						Version: "1.0.0",
 					},
 				},
@@ -163,7 +163,7 @@ func TestPublishEndpoint(t *testing.T) {
 				Server: model.ServerDetail{
 					Name:        "test-server",
 					Description: "A test server",
-					VersionDetail: model.VersionDetail{
+					VersionDetail: pkgmodel.VersionDetail{
 						Version: "1.0.0",
 					},
 				},
@@ -179,7 +179,7 @@ func TestPublishEndpoint(t *testing.T) {
 				Server: model.ServerDetail{
 					Name:        "io.github.other/test-server",
 					Description: "A test server",
-					VersionDetail: model.VersionDetail{
+					VersionDetail: pkgmodel.VersionDetail{
 						Version: "1.0.0",
 					},
 					Repository: pkgmodel.Repository{
@@ -205,7 +205,7 @@ func TestPublishEndpoint(t *testing.T) {
 				Server: model.ServerDetail{
 					Name:        "example/test-server",
 					Description: "A test server",
-					VersionDetail: model.VersionDetail{
+					VersionDetail: pkgmodel.VersionDetail{
 						Version: "1.0.0",
 					},
 					Repository: pkgmodel.Repository{
