+++ internal/model/model.go
@@ -11,58 +11,6 @@ import (
 	pkgmodel "github.com/modelcontextprotocol/registry/pkg/model"
 )
 
-type InputWithVariables struct {
-	pkgmodel.Input `json:",inline" bson:",inline"`
-	Variables      map[string]pkgmodel.Input `json:"variables,omitempty" bson:"variables,omitempty"`
-}
-
-type KeyValueInput struct {
-	InputWithVariables `json:",inline" bson:",inline"`
-	Name               string `json:"name" bson:"name"`
-}
-type ArgumentType string
-
-const (
-	ArgumentTypePositional ArgumentType = "positional"
-	ArgumentTypeNamed      ArgumentType = "named"
-)
-
-// RuntimeArgument defines a type that can be either a PositionalArgument or a NamedArgument
-type Argument struct {
-	InputWithVariables `json:",inline" bson:",inline"`
-	Type               ArgumentType `json:"type" bson:"type"`
-	Name               string       `json:"name,omitempty" bson:"name,omitempty"`
-	IsRepeated         bool         `json:"is_repeated,omitempty" bson:"is_repeated,omitempty"`
-	ValueHint          string       `json:"value_hint,omitempty" bson:"value_hint,omitempty"`
-}
-
-type Package struct {
-	// RegistryType indicates how to download packages (e.g., "npm", "pypi", "docker-hub", "mcpb")
-	RegistryType string `json:"registry_type,omitempty" bson:"registry_type,omitempty"`
-	// RegistryBaseURL is the base URL of the package registry
-	RegistryBaseURL string `json:"registry_base_url,omitempty" bson:"registry_base_url,omitempty"`
-	// Identifier is the package identifier - either a package name (for registries) or URL (for direct downloads)
-	Identifier           string          `json:"identifier,omitempty" bson:"identifier,omitempty"`
-	Version              string          `json:"version,omitempty" bson:"version,omitempty"`
-	FileSHA256           string          `json:"file_sha256,omitempty" bson:"file_sha256,omitempty"`
-	RunTimeHint          string          `json:"runtime_hint,omitempty" bson:"runtime_hint,omitempty"`
-	RuntimeArguments     []Argument      `json:"runtime_arguments,omitempty" bson:"runtime_arguments,omitempty"`
-	PackageArguments     []Argument      `json:"package_arguments,omitempty" bson:"package_arguments,omitempty"`
-	EnvironmentVariables []KeyValueInput `json:"environment_variables,omitempty" bson:"environment_variables,omitempty"`
-}
-
-// Remote represents a remote connection endpoint
-type Remote struct {
-	TransportType string          `json:"transport_type" bson:"transport_type"`
-	URL           string          `json:"url" format:"uri" bson:"url"`
-	Headers       []KeyValueInput `json:"headers,omitempty" bson:"headers,omitempty"`
-}
-
-// VersionDetail represents the version details of a server (pure MCP spec, no registry metadata)
-type VersionDetail struct {
-	Version string `json:"version" bson:"version"`
-}
-
 // ServerDetail represents complete server information as defined in the MCP spec (pure, no registry metadata)
 type ServerDetail struct {
 	Schema        string                `json:"$schema,omitempty" bson:"$schema,omitempty"`
@@ -70,9 +18,9 @@ type ServerDetail struct {
 	Description   string                `json:"description" minLength:"1" maxLength:"100" bson:"description"`
 	Status        pkgmodel.ServerStatus `json:"status,omitempty" minLength:"1" bson:"status,omitempty"`
 	Repository    pkgmodel.Repository   `json:"repository,omitempty" bson:"repository"`
-	VersionDetail VersionDetail         `json:"version_detail" bson:"version_detail"`
-	Packages      []Package             `json:"packages,omitempty" bson:"packages,omitempty"`
-	Remotes       []Remote              `json:"remotes,omitempty" bson:"remotes,omitempty"`
+	VersionDetail pkgmodel.VersionDetail `json:"version_detail" bson:"version_detail"`
+	Packages      []pkgmodel.Package    `json:"packages,omitempty" bson:"packages,omitempty"`
+	Remotes       []pkgmodel.Remote     `json:"remotes,omitempty" bson:"remotes,omitempty"`
 }
 
 // RegistryMetadata represents registry-generated metadata
