+++ internal/service/registry_service_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/modelcontextprotocol/registry/internal/database"
 	"github.com/modelcontextprotocol/registry/internal/model"
+	pkgmodel "github.com/modelcontextprotocol/registry/pkg/model"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -16,10 +17,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 		"existing1": {
 			Name:        "com.example/existing-server",
 			Description: "An existing server",
-			VersionDetail: model.VersionDetail{
+			VersionDetail: pkgmodel.VersionDetail{
 				Version: "1.0.0",
 			},
-			Remotes: []model.Remote{
+			Remotes: []pkgmodel.Remote{
 				{URL: "https://api.example.com/mcp"},
 				{URL: "https://webhook.example.com/sse"},
 			},
@@ -27,10 +28,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 		"existing2": {
 			Name:        "com.microsoft/another-server",
 			Description: "Another existing server",
-			VersionDetail: model.VersionDetail{
+			VersionDetail: pkgmodel.VersionDetail{
 				Version: "1.0.0",
 			},
-			Remotes: []model.Remote{
+			Remotes: []pkgmodel.Remote{
 				{URL: "https://api.microsoft.com/mcp"},
 			},
 		},
@@ -50,10 +51,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 			serverDetail: model.ServerDetail{
 				Name:        "com.example/new-server",
 				Description: "A new server with no remotes",
-				VersionDetail: model.VersionDetail{
+				VersionDetail: pkgmodel.VersionDetail{
 					Version: "1.0.0",
 				},
-				Remotes: []model.Remote{},
+				Remotes: []pkgmodel.Remote{},
 			},
 			expectError: false,
 		},
@@ -62,10 +63,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 			serverDetail: model.ServerDetail{
 				Name:        "com.example/new-server",
 				Description: "A new server",
-				VersionDetail: model.VersionDetail{
+				VersionDetail: pkgmodel.VersionDetail{
 					Version: "1.0.0",
 				},
-				Remotes: []model.Remote{
+				Remotes: []pkgmodel.Remote{
 					{URL: "https://new.example.com/mcp"},
 					{URL: "https://unique.example.com/sse"},
 				},
@@ -77,10 +78,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 			serverDetail: model.ServerDetail{
 				Name:        "com.example/new-server",
 				Description: "A new server with duplicate URL",
-				VersionDetail: model.VersionDetail{
+				VersionDetail: pkgmodel.VersionDetail{
 					Version: "1.0.0",
 				},
-				Remotes: []model.Remote{
+				Remotes: []pkgmodel.Remote{
 					{URL: "https://api.example.com/mcp"}, // This URL already exists
 				},
 			},
@@ -92,10 +93,10 @@ func TestValidateNoDuplicateRemoteURLs(t *testing.T) {
 			serverDetail: model.ServerDetail{
 				Name:        "com.example/existing-server", // Same name as existing
 				Description: "Updated existing server",
-				VersionDetail: model.VersionDetail{
+				VersionDetail: pkgmodel.VersionDetail{
 					Version: "1.1.0",
 				},
-				Remotes: []model.Remote{
+				Remotes: []pkgmodel.Remote{
 					{URL: "https://api.example.com/mcp"}, // Same URL as before
 				},
 			},
